requirements.txt
pymongo>=4.5.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0
pytz>=2023.3
python-dateutil>=2.8.2 